🎨 Updated startGame() with printQuestion()
Removed separate event-listener for playAgainBtn in main and exported the printQUestions() to use in the startGame() function instead.

Do not know if it is best practise to export from main to then import a function with that export back to main, but it seems to work so maybe it is ok. Otherwise we can go back to a separate wventlistener instead. :)